+++ plantio/models/__init__.py
@@ -1,1 +1,2 @@
 from .plantio import *  # noqa
+from .aplicacao import *  # noqa
