+++ plantio/urls.py
@@ -1,14 +1,13 @@
 from django.conf.urls import url
 from django.urls import include
 
-from plantio.views.plantio import PlantioAPIView, PlantioAssociacaoAPIView
+from plantio.views.plantio import PlantioAPIView
+from plantio.views.aplicacao import AplicacaoAgrotoxicoAPIView
 
 
 urlpatterns = [
     url(r'^plantio/', include([
         url(r'^$', PlantioAPIView.as_view(), name='plantio-create'),
-        url(
-            r'^(?P<idPlantio>[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})/$',
-            PlantioAssociacaoAPIView.as_view(), name='plantio-associar'),
+        url(r'^aplicar/$', AplicacaoAgrotoxicoAPIView.as_view(), name='plantio-associar'),
     ]))
 ]
